chore(docs): standardise image paths for both GitHub and Jekyll compatibility
This commit attempts all images being properly displayed when accessed in both the GitHub web interface, as in GitHub Pages (Jekyll). So far images were only linked successfully when accessed via the GitHub web interface, but not in GitHub Pages. Important to mention thought: this change is a preliminary first test.

- Added `baseurl: "/zephyr-rtos-tutorial"` to `_config.yml` to set a base path configuration for Jekyll.
- Updated all image paths in the documentation to use `{{ site.baseurl }}` instead of `../../` parent references.